requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
torch>=2.0.0
transformers>=4.21.0
nltk>=3.8
PyPDF2>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Enhanced PDF processing
PyMuPDF>=1.23.0
Pillow>=10.0.0

# OCR support for scanned PDFs
pytesseract>=0.3.10
opencv-python>=4.8.0

# Audio processing (multimodal support)
librosa>=0.10.0
pydub>=0.25.1
openai-whisper>=20231117
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Enhanced text processing
python-docx>=1.1.0
ebooklib>=0.18
html2text>=2020.1.16